test: add bench tests in db_dbtvec.rs
Changes:
 * make some bench tests for dbtvec read/write ops.
 * rename bench/database.rs to db_leveldb.rs
